+++ src/ideabox/policy/__init__.py
@@ -2,5 +2,8 @@
 
 from zope.i18nmessageid import MessageFactory
 
+import logging
+
 
 _ = MessageFactory('ideabox.policy')
+logger = logging.getLogger('ideabox.policy')